slightly simplify codebase
